requirements.txt
streamlit
pathlib
pandas
pillow